requirements.txt
fastapi
python-jose
psycopg2-binary
sqlalchemy
pg8000